use u32 as the penalty and length instead of usize

+++ sigalign/src/core/regulators.rs
@@ -1,22 +1,23 @@
-/// Alignment conditions that affect the alignment result
+/// Alignment regulators
+///   - Type is determined by machine's pointer width
 
-pub const PREC_SCALE: usize = 100_000; // Ensuring accuracy to the fourth decimal place.
+pub const PREC_SCALE: u32 = 100_000; // Ensuring accuracy to the fourth decimal place.
 
 #[derive(Debug, Clone, PartialEq, Eq, PartialOrd, Ord)]
 pub struct Penalty {
-    pub x: usize,
-    pub o: usize,
-    pub e: usize,
+    pub x: u32,
+    pub o: u32,
+    pub e: u32,
 }
 
 #[derive(Debug, Clone, PartialEq, Eq, PartialOrd, Ord)]
 pub struct Cutoff {
-    pub minimum_aligned_length: usize,
-    pub maximum_penalty_per_scale: usize,
+    pub minimum_aligned_length: u32,
+    pub maximum_penalty_per_scale: u32,
 }
 
 #[derive(Debug, Clone, PartialEq, Eq, PartialOrd, Ord)]
 pub struct MinPenaltyForPattern {
-    pub odd: usize,
-    pub even: usize,
+    pub odd: u32,
+    pub even: u32,
 }
